feat: Code rework for better testing and mocking
